requirements.txt
Click==7.0
docopt==0.6.2
dominate==2.3.5
Flask==1.0.2
Flask-Bootstrap==3.3.7.1
Flask-Debug==0.4.3
flask-nav==0.6
Flask-Sessionstore==0.4.5
Flask-WTF==0.14.2
idna==2.8
inflection==0.3.1
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
monotonic==1.5
python-valve==0.2.1
requests==2.21.0
six==1.12.0
urllib3==1.25
visitor==0.1.3
Werkzeug==0.15.2
WTForms==2.2.1
